top-level: Introduce `buildPackages` for resolving  build-time deps
In most cases, buildPackages = pkgs so things work just as before. For
cross compiling, however, buildPackages is resolved as the previous
bootstrapping stage. This allows us to avoid the mkDerivation hacks cross
compiling currently uses today.

To avoid a massive refactor, callPackage will splice together both package
sets. Again to avoid churn, it uses the old `nativeDrv` vs `crossDrv` to do
so. So now, whether cross compiling or not, packages with get a `nativeDrv`
and `crossDrv`---in the non-cross-compiling case they are simply the same
derivation. This is good because it reduces the divergence between the
cross and non-cross dataflow. See `pkgs/top-level/splice.nix` for a comment
along the lines of the preceding paragraph, and the code that does this
splicing.

With cross compiling now using proper bootstrapping stages, we can start
cleaning things up (though most cleanup I leave for future commits).
`stdenv.cross` was used informally before as the target platform, it now
should consistently be used as that, and `crossPlatform` relegated to mean
the host platform.

Also, `forceNativeDrv` is replaced with `forceNativePackages`. The latter
resolves `pkgs` unless the host platform is different from the build
platform, in which case it resolves to `buildPackages`. Note that the
target platform is not important here---it will not prevent
`forcedNativePackages` from resolving to `pkgs`.

--------

Temporarily, we make some dubious decisions in the name of preserving
hashes:

Most importantly, we don't resolve native deps of the cross "build
packages" (build = host != target) package set against the "vanilla
packages" (build = host = target) package set. This is wrong because it
violates the "sliding window" principle of bootstrapping stages that
shifting the platform triple of one stage to the left coincides with the
next stage's platform triple.

Additionally, we build libc, libgcc, etc in the same stage as the compilers
themselves, which is wrong because they are used at runtime, not build
time. Fixing this is somewhat subtle, and the solution and problem will be
better explained in the commit that does fix it.

Commits after this will solve both these issues, at the expense of breaking
cross hashes. Native hashes won't be broken, thankfully.

--------

Despite my efforts, not all cross hashes are preserved. On can do
```sh
printf ":p import pkgs/top-level/release-cross.nix\n:q" | nix-repl | sed 's/»/»\n/g' > cross.txt
```
before and after this commit to see the fallout. (`nix-repl` pretty-prints
in the presence of assertion failures better, hence why I use it not
`nix-instantiate`.)

So far, it seems this is mainly due to the old cross compiling
infrastructure not properly handling multiple outputs. For example, apply
the following *before* this commit and then do the same experiment for a
slight improvement.
```diff
--- a/pkgs/applications/networking/browsers/lynx/default.nix
+++ b/pkgs/applications/networking/browsers/lynx/default.nix
@@ -14,7 +14,7 @@ stdenv.mkDerivation rec {
   };

   configureFlags = []
-    ++ stdenv.lib.optionals sslSupport [ "--with-ssl=${openssl.dev}" ];
+    ++ stdenv.lib.optionals sslSupport [ "--with-ssl=${(openssl.crossDrv or openssl).dev}" ];

   buildInputs = [ ncurses gzip ];
   nativeBuildInputs = [ ncurses ];
--- a/pkgs/tools/compression/gzip/default.nix
+++ b/pkgs/tools/compression/gzip/default.nix
@@ -13,7 +13,7 @@ stdenv.mkDerivation rec {

   enableParallelBuilding = true;

-  buildInputs = [ xz.bin ];
+  buildInputs = [ (xz.crossDrv or xz).bin ];

   preConfigure = if stdenv.isCygwin then ''
     sed -i lib/fpending.h -e 's,include <stdio_ext.h>,,'
```

Since the following commits will purposefully break hashes anyways, I
consider the current state of breakage good enough.