bintools-wrapper: Don't bother with `libc.bin`
`ldd` I suppose is a binary tool, but we have no reason to wrap it and
anyone that wants it can depend on `libc.bin` directly.

I don't think this will be a be big breaking change in practice becausee
cc-wrapper still provides it, and the vast majority of uses of
bintools-wrapper are via that.